Treinamento quarkus e kafka turma noite

+++ src/main/java/com/alura/agencias/domain/Agencia.java
@@ -9,7 +9,7 @@ public class Agencia {
 
     }
 
-    public Agencia(Integer id, String nome, String razaoSocial, String cnpj, Endereco endereco) {
+    public Agencia(Long id, String nome, String razaoSocial, String cnpj, Endereco endereco) {
         this.id = id;
         this.nome = nome;
         this.razaoSocial = razaoSocial;
@@ -19,7 +19,7 @@ public class Agencia {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
 
     @Column(name = "razao_social")
@@ -30,7 +30,7 @@ public class Agencia {
     @JoinColumn(name = "endereco_id")
     private Endereco endereco;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
